machine_b/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pyyaml>=6.0.1
